+++ NEWS.md
@@ -1,3 +1,8 @@
+## 12.5.0 Unreleased
+
+* Add `service-points-users` endpoint (MODINVSTOR-132)
+* Provides `service-points-users` interface 1.0 (MODINVSTOR-132)
+
 ## 12.4.0 2018-07-26
 
 * Item `status` now defaults to 'Available' (MODINVSTOR-137)